requirements.txt
faster-whisper
silero-vad
sounddevice
numpy
pynput
pyperclip
torch
torchaudio
pystray
Pillow
